feat: :truck: Organised the functions into seperate files and referenced them in the main file
Solved the other issues pointed in the pr like output file, tsdoc changes, new line before }
