+++ src/parser/inner_type_from_path_segment.rs
@@ -1,0 +1,82 @@
+use syn::Type;
+
+pub fn inner_type_from_path_segment(segment: &syn::PathSegment) -> Option<String> {
+    if segment.ident == "Result" {
+        match &segment.arguments {
+            syn::PathArguments::AngleBracketed(params) => {
+                if let Some(syn::GenericArgument::Type(Type::Path(inner_type))) =
+                    params.args.first()
+                {
+                    if let Some(inner_segment) = inner_type.path.segments.first() {
+                        if inner_segment.ident == "Json" {
+                            if let syn::PathArguments::AngleBracketed(inner_params) =
+                                &inner_segment.arguments
+                            {
+                                if let Some(syn::GenericArgument::Type(Type::Path(
+                                    most_inner_type,
+                                ))) = inner_params.args.first()
+                                {
+                                    return Some(
+                                        most_inner_type
+                                            .path
+                                            .segments
+                                            .last()
+                                            .unwrap()
+                                            .ident
+                                            .to_string(),
+                                    );
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+            syn::PathArguments::Parenthesized(_params) => {
+                unimplemented!()
+            }
+            syn::PathArguments::None => return Some(String::from("void")),
+        }
+    } else if segment.ident == "Json" {
+        // println!("segment.ident =  Json");
+
+        match &segment.arguments {
+            syn::PathArguments::AngleBracketed(params) => {
+                if let Some(syn::GenericArgument::Type(Type::Path(inner_type))) =
+                    params.args.first()
+                {
+                    if let Some(inner_segment) = inner_type.path.segments.first() {
+                        if let syn::PathArguments::AngleBracketed(inner_params) =
+                            &inner_segment.arguments
+                        {
+                            if let Some(syn::GenericArgument::Type(Type::Path(most_inner_type))) =
+                                inner_params.args.first()
+                            {
+                                return Some(
+                                    most_inner_type
+                                        .path
+                                        .segments
+                                        .last()
+                                        .unwrap()
+                                        .ident
+                                        .to_string(),
+                                );
+                            }
+                        } else {
+                            return Some(inner_segment.ident.to_string());
+                        }
+                    }
+                }
+            }
+            syn::PathArguments::Parenthesized(_params) => {
+                unimplemented!()
+            }
+            syn::PathArguments::None => return Some(String::from("void")),
+        }
+    } else {
+        // println!("segment.ident = {}", segment.ident);
+    }
+
+    // println!("inner_type_from_path_segment ending w/ None");
+
+    None
+}
